requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
langchain-openai>=0.0.5
langchain-community>=0.0.16
openai>=1.10.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
requests>=2.31.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
pillow>=10.0.0
opencv-python>=4.7.0 